+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
  - Delete stalled spams if UserNotFound
  - Added more info in logs (guild and channel ID)
  - Removed ambiguous logs when resuming spams
+ - `help` command only shows available commands
 
 ## [0.0.5.2]